Adds `retrieve_metadata` function to axom's annotations API
This returns the registered metadata a key-value map of strings
and was previously implemented in the annotations unit tests.
